convertDisplayDialySalesをユーティリティとして別ファイルに切り出し
