+++ src/components/DialySales.tsx
@@ -5,7 +5,7 @@ import { DataGrid, GridColDef } from "@mui/x-data-grid";
 import { DialySalesStateContext } from "../context/DialySalesContext";
 import { SelectDialySalesContext } from "../context/SelectDialySalesContext";
 import { LOCAL_DIALYSALES_ADDRESS } from "@/constants/serverAdress";
-import dayjs from "dayjs";
+import { convertDisplayDialySales } from "@/util/convertDisplayDialySales";
 
 export type DialySalesProps = {
   firstDialySaleDay?: string;
@@ -18,31 +18,6 @@ export const DialySales = (props: DialySalesProps) => {
 
   const { state, dispatch } = useContext(DialySalesStateContext);
   const { setRowSelectionModel } = useContext(SelectDialySalesContext);
-
-  //サーバーから取得したISO8601規格のsalesDayをDialySalesでの表示形式に変換して返却する
-  const salesDayFormatToDisplay = (_salesDay: string): string => {
-    const salesDayToDate = dayjs(_salesDay);
-    return salesDayToDate.format("YYYY/MM/DD(ddd)");
-  };
-
-  //サーバーから取得した値を画面表示形式に変換する
-  const convertDisplayDialySales = (fetchData: GetFromSeverDialySale[]): DisplayDialySale[] => {
-    const displayDialySales = fetchData.map((data) => {
-      const convertFetchData: DisplayDialySale = {} as DisplayDialySale;
-      convertFetchData.id = data.id;
-      convertFetchData.salesDay = salesDayFormatToDisplay(data.sales_day);
-      convertFetchData.lunchSales = data.lunch_sales;
-      convertFetchData.dinnerSales = data.dinner_sales;
-      convertFetchData.lunchVisitor = data.lunch_visitor;
-      convertFetchData.dinnerVisitor = data.dinner_visitor;
-      convertFetchData.personnelCost = data.personnel_cost;
-      convertFetchData.purchase = data.purchase;
-      convertFetchData.totalSale = data.lunch_sales + data.dinner_sales;
-      convertFetchData.totalVisitor = data.lunch_visitor + data.dinner_visitor;
-      return convertFetchData;
-    });
-    return displayDialySales;
-  };
 
   //DialySale一覧を取得する関数
   const fetchDialySales = async () => {
